+++ CHANGELOG.md
@@ -5,6 +5,15 @@ is based on https://keepachangelog.com/en/1.1.0/.
 
 ## Unreleased
 
+### Added
+
+- `Display` impl for `instruction::Instruction`, `instruction::Kind` and the
+  various `instruction::format::Type*`.
+
+### Changed
+
+- Include (known) disassembled instruction in output of simple example.
+
 ## 0.2.0 - 2025-07-16
 
 ### Added